+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { Provider as JotaiProvider, useAtom } from 'jotai';
+import { useAtom } from 'jotai';
 import Router from './shared/Router';
-import { themeAtom, backgroundImageAtom } from './atoms/ThemaAtom';
+import { themeAtom, backgroundImageAtom } from './atoms/Atom';
 import { userAtom } from './atoms/Atom';
 import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from './firebase/firebaseConfig';
@@ -22,7 +22,7 @@ function App() {
 
     // cleanup 함수 등록
     return () => unsubscribe();
-  }, [setUser]);
+  }, []);
 
   useEffect(() => {
     // 로컬 스토리지에서 저장된 테마 정보 불러오기
@@ -55,9 +55,7 @@ function App() {
 
   return (
     <QueryClientProvider client={queryClient}>
-      <JotaiProvider>
-        <Router />
-      </JotaiProvider>
+      <Router />
     </QueryClientProvider>
   );
 }
